feat: swagger init
