+++ server/handlers/authorHandler/schemas.js
@@ -1,0 +1,9 @@
+const Joi = require("joi");
+
+const authorGetQuerySchema = Joi.object({
+  id: Joi.string().uuid(),
+}).unknown(false);
+
+module.exports = {
+  authorGetQuerySchema,
+};
