+++ server/handlers/surveyHandler/schemas.js
@@ -1,0 +1,9 @@
+const Joi = require("joi");
+
+const surveyIdParamSchema = Joi.object({
+  uuid: Joi.string().required(),
+}).unknown(false);
+
+module.exports = {
+  surveyIdParamSchema,
+};
